refactor: move SystemEnv out of class
