+++ app/src/main/java/org/puffinbasic/runtime/Environment.kt
@@ -3,20 +3,21 @@ package org.puffinbasic.runtime
 interface Environment {
     operator fun get(key: String): String
     operator fun set(key: String, value: String)
-    class SystemEnv : Environment {
-        private val overrides: MutableMap<String, String>
+}
 
-        init {
-            overrides = HashMap()
-        }
+class SystemEnv : Environment {
+    private val overrides: MutableMap<String, String>
 
-        override fun get(key: String): String {
-            val result = overrides[key]
-            return result ?: System.getenv(key)
-        }
+    init {
+        overrides = HashMap()
+    }
 
-        override fun set(key: String, value: String) {
-            overrides[key] = value
-        }
+    override fun get(key: String): String {
+        val result = overrides[key]
+        return result ?: System.getenv(key) ?: ""
+    }
+
+    override fun set(key: String, value: String) {
+        overrides[key] = value
     }
 }
